@rickar 2020-08-24
Nice, thanks! I actually considered creating a cache of the IsHoliday results in Calendar for the last few used years with the assumption that the same 2-3 years will be called multiple times in a row for normal usage. But it would make the code more complicated so I thought I would wait until it was actually needed. Maybe that's the next thing to look at if the performance isn't good enough for you.